+++ modules/pcd/src/lib/parse-pcd.ts
@@ -8,6 +8,7 @@
 
 import type {MeshAttribute, MeshAttributes} from '@loaders.gl/schema';
 import {getMeshBoundingBox} from '@loaders.gl/schema';
+import {decompressLZF} from './decompress-lzf';
 import {getPCDSchema} from './get-pcd-schema';
 import type {PCDHeader} from './pcd-types';
 
@@ -316,12 +317,14 @@ function parsePCDBinary(pcdHeader: PCDHeader, data: ArrayBufferLike): HeaderAttr
   return {position, normal, color};
 }
 
-/**
+/** Parse compressed PCD data in in binary_compressed form ( https://pointclouds.org/documentation/tutorials/pcd_file_format.html)
+ * from https://github.com/mrdoob/three.js/blob/master/examples/jsm/loaders/PCDLoader.js
+ * @license MIT (http://opensource.org/licenses/MIT)
  * @param pcdHeader
  * @param data
  * @returns [attributes]
  */
-function parsePCDBinaryCompressed(PCDheader, data) {
+function parsePCDBinaryCompressed(PCDheader: PCDHeader, data: ArrayBufferLike): HeaderAttributes {
   const position: number[] = [];
   const normal: number[] = [];
   const color: number[] = [];
@@ -340,44 +343,59 @@ function parsePCDBinaryCompressed(PCDheader, data) {
   for (let i = 0; i < PCDheader.points; i++) {
     if (offset.x !== undefined) {
       position.push(
-        dataview.getFloat32(PCDheader.points * offset.x + PCDheader.size[0] * i, LITTLE_ENDIAN)
+        dataview.getFloat32(
+          (PCDheader.points as number) * offset.x + (PCDheader.size as number[])[0] * i,
+          LITTLE_ENDIAN
+        )
       );
       position.push(
-        dataview.getFloat32(PCDheader.points * offset.y + PCDheader.size[1] * i, LITTLE_ENDIAN)
+        dataview.getFloat32(
+          (PCDheader.points as number) * offset.y + (PCDheader.size as number[])[1] * i,
+          LITTLE_ENDIAN
+        )
       );
       position.push(
-        dataview.getFloat32(PCDheader.points * offset.z + PCDheader.size[2] * i, LITTLE_ENDIAN)
+        dataview.getFloat32(
+          (PCDheader.points as number) * offset.z + (PCDheader.size as number[])[2] * i,
+          LITTLE_ENDIAN
+        )
       );
     }
 
     if (offset.rgb !== undefined) {
       color.push(
-        dataview.getUint8(PCDheader.points * offset.rgb + PCDheader.size[3] * i + 0) / 255.0
+        dataview.getUint8(
+          (PCDheader.points as number) * offset.rgb + (PCDheader.size as number[])[3] * i + 0
+        ) / 255.0
       );
       color.push(
-        dataview.getUint8(PCDheader.points * offset.rgb + PCDheader.size[3] * i + 1) / 255.0
+        dataview.getUint8(
+          (PCDheader.points as number) * offset.rgb + (PCDheader.size as number[])[3] * i + 1
+        ) / 255.0
       );
       color.push(
-        dataview.getUint8(PCDheader.points * offset.rgb + PCDheader.size[3] * i + 2) / 255.0
+        dataview.getUint8(
+          (PCDheader.points as number) * offset.rgb + (PCDheader.size as number[])[3] * i + 2
+        ) / 255.0
       );
     }
 
     if (offset.normal_x !== undefined) {
       normal.push(
         dataview.getFloat32(
-          PCDheader.points * offset.normal_x + PCDheader.size[4] * i,
+          (PCDheader.points as number) * offset.normal_x + (PCDheader.size as number[])[4] * i,
           LITTLE_ENDIAN
         )
       );
       normal.push(
         dataview.getFloat32(
-          PCDheader.points * offset.normal_y + PCDheader.size[5] * i,
+          (PCDheader.points as number) * offset.normal_y + (PCDheader.size as number[])[5] * i,
           LITTLE_ENDIAN
         )
       );
       normal.push(
         dataview.getFloat32(
-          PCDheader.points * offset.normal_z + PCDheader.size[6] * i,
+          (PCDheader.points as number) * offset.normal_z + (PCDheader.size as number[])[6] * i,
           LITTLE_ENDIAN
         )
       );
@@ -389,48 +407,4 @@ function parsePCDBinaryCompressed(PCDheader, data) {
     normal,
     color
   };
-}
-
-function decompressLZF(inData, outLength) {
-  const inLength = inData.length;
-  const outData = new Uint8Array(outLength);
-  let inPtr = 0;
-  let outPtr = 0;
-  let ctrl;
-  let len;
-  let ref;
-
-  do {
-    ctrl = inData[inPtr++];
-
-    if (ctrl < 1 << 5) {
-      ctrl++;
-      if (outPtr + ctrl > outLength) throw new Error('Output buffer is not large enough');
-      if (inPtr + ctrl > inLength) throw new Error('Invalid compressed data');
-
-      do {
-        outData[outPtr++] = inData[inPtr++];
-      } while (--ctrl);
-    } else {
-      len = ctrl >> 5;
-      ref = outPtr - ((ctrl & 0x1f) << 8) - 1;
-      if (inPtr >= inLength) throw new Error('Invalid compressed data');
-
-      if (len === 7) {
-        len += inData[inPtr++];
-        if (inPtr >= inLength) throw new Error('Invalid compressed data');
-      }
-
-      ref -= inData[inPtr++];
-      if (outPtr + len + 2 > outLength) throw new Error('Output buffer is not large enough');
-      if (ref < 0) throw new Error('Invalid compressed data');
-      if (ref >= outPtr) throw new Error('Invalid compressed data');
-
-      do {
-        outData[outPtr++] = outData[ref++];
-      } while (--len + 2);
-    }
-  } while (inPtr < inLength);
-
-  return outData;
 }