+++ docs/whats-new.md
@@ -59,6 +59,9 @@ Target Release Date: Nov 2021.
 
 - New experimental [`ParquetLoader`](modules/parquet/docs/api-reference/parquet-loader)
 
+**@loaders.gl/pcd**
+- `PCDLoader` now supports for compressed binary data.
+
 ## v3.0
 
 Release Date: July 13, 2021.